+++ source/Application/SearchMessages/GetMessageQueryHandler.cs
@@ -1,0 +1,37 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Domain.ArchivedMessages;
+using MediatR;
+
+namespace Application.SearchMessages;
+
+public class GetMessageQueryHandler : IRequestHandler<GetMessagesQuery, MessageSearchResult>
+{
+    private readonly IArchivedMessageRepository _archivedMessageRepository;
+
+    public GetMessageQueryHandler(IArchivedMessageRepository archivedMessageRepository)
+    {
+        _archivedMessageRepository = archivedMessageRepository;
+    }
+
+    public async Task<MessageSearchResult> Handle(GetMessagesQuery request, CancellationToken cancellationToken)
+    {
+        var messages = await _archivedMessageRepository.GetAllAsync().ConfigureAwait(false);
+        return new MessageSearchResult(messages.Select(message => new MessageInfo(message.MessageId)).ToList().AsReadOnly());
+    }
+}
