+++ source/IntegrationTests/TestBase.cs
@@ -23,6 +23,7 @@ using Application.Configuration.Commands.Commands;
 using Application.Configuration.DataAccess;
 using Application.Configuration.Queries;
 using Application.Configuration.TimeEvents;
+using Application.SearchMessages;
 using Application.Transactions.MoveIn;
 using Azure.Messaging.ServiceBus;
 using Google.Protobuf;
